Refactor: 좌표값 정리 완료 #15
1. 좌표값 소수점 안되는거 정리완료

+++ src/main/java/com/solver/solver_be/domain/company/dto/CompanyRequestDto.java
@@ -11,8 +11,8 @@ public class CompanyRequestDto {
 
     private String companyName;
 
-    private Integer x;
+    private Double x;
 
-    private Integer y;
+    private Double y;
 
 }
